pyproject.toml
[tool.poetry]
name = "open-assistant-api"
version = "0.1.0"
description = ""
authors = ["李猛 <limeng.b@mininglamp.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "default"

[tool.black]
target-version = ["py310"]
line-length = 120

[tool.ruff]
exclude = ["migrations", "tests/e2e"]
# https://docs.astral.sh/ruff/rules
select = ["E", "F", "N", "PIE", "T20", "Q"]
line-length = 120


[tool.poetry.dependencies]
python = "^3.10"
celery = ">=5.3.4,<5.4.0"
fastapi = ">=0.94.1,<0.95.0"
fastapi-pagination = ">=0.12.12,<0.13.0"
sqlakeyset = ">=2.0.1680321678,<2.1.0"
loguru = ">=0.6.0,<0.7.0"
sqlmodel = ">=0.0.11,<0.1.0"
starlette = ">=0.26.1,<0.27.0"
uvicorn = ">=0.21.0,<0.22.0"
pymysql = ">=1.1.0,<1.2.0"
python-multipart = ">=0.0.6,<0.1.0"
redis = ">=5.0.1,<5.1.0"
orjson = ">=3.9.10,<3.10.0"
boto3 = ">=1.33.0,<1.34.0"
langchain = ">=0.0.342,<0.1.0"
openai = ">=1.3.9,<1.4.0"
python-magic = ">=0.4.27,<0.5.0"
beautifulsoup4 = ">=4.12.2,<4.13.0"
pymupdf = ">=1.23.7,<1.24.0"
alembic = ">=1.13.0,<1.14.0"
pydantic = { version = ">=1.10.13,<1.11.0", extras = ["dotenv"] }


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
ruff = "^0.1.9"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
